[ST40] Check parameters when using the RET_xxx() re-timing macros.
This commit forces all the RET_xxx() re-timing macros, which are used to
initialize an instance of the "stm_pio_control_retime_config" structure,
to use "valid" values only. Specifically, for the "clk" and the "delay"
(multi-bit) bitfields.

If an attempt is made to pass in invalid parameter to the RET_xxx()
macros, then a compilation error will now result. This is particularly
useful now that the sizes of the above two bitfields (in H/W) can vary
depending on the actual SoC being used. Also, the semantics of the
"delay" bitfield in the linux kernel sources are now in units of ps
(pico-seconds), but in U-Boot it is just the raw binary value to be
programmed. This should avoid users unwittingly copying code from linux,
to configure the re-timings registers, and the compiler silently ignoring
the paradigm shift (in semantics) in U-Boot.

This has not been tested, apart from compilation.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>