[ST40] Support for the "DELAY_INNOTOUT" bitfield on STxH416 (Orly2)
This commit adds (the missing) support for the "DELAY_INNOTOUT" bitfield
(position #7) in the PIO re-time configuration register space, on the
STMicroelectronics' STxH416 (Orly2) SoC.

The STxH416 (Orly2) SoC introduces a new bit-field in the re-time
configuration registers, which controls if the additional delay is to be
added on the input, or the output side. This commit essentially extends
the API of the function stxh416_pioalt_retime() to allow us to pass in
an additional parameter to include the PIO's directionality, and use
that to set the aforementioned DELAY_INNOTOUT bitfield accordingly.

This has not been tested, apart from compilation.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>