[ST40] Optimize stxh416_pioalt_retime()
This is a minor update to the function stxh416_pioalt_retime() for two
main reasons:

	1) Optimize the code slightly, by reading/writing the appropriate
	   (volatile) system configuration register only once per
	   invocation (instead of once per bit-field per invocation).

	2) Re-order (for maintainability reasons) the code to put the
	   bit-twiddling operations in (increasing) order of the bit-field
	   positions (lowest to highest).

Note: the code will still honour the "don't care" semantics, if the
values to be written in the structure are negative (-1). However, this
usage is deprecated, and will be removed.

This has not been tested, apart from compilation.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>