[ST40] Add support for the 4th USB port on the STxH416 (Orly2) SoC
This commit adds support for the 4th USB (USB #3) on the STxH416 (Orly2)
SoC from STMicroelectronics. Previously only the first three (0-2)
were usable.

As a consequence of this, we also had to add support for PIO port #40.
We also corrected a bug that USB[x]_POWERDOWN_ACK is actually in system
config register #2583, and not system config register #384.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>