[STM] Dump out the ARM ATAGS to facilitate debugging
This commit adds some (conditional) debug code, to dump out the raw ARM
ATAGS data structures (as 32-bit words), before launching the ARM linux
kernel, from U-Boot running on the ST40 core.

The data is dumped simply as raw 32-bit hexadecimal values, which gives
some useful visibility, if the ARM kernel is not booting correctly. In the
future a more sophisticated version may "decode" these data structures,
and print it an easier to read format.

To enable this new capability, the following #define needs to be defined:

	#define CONFIG_STM_DUMP_ATAGS

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>