[STM] Completed initial testing of B2000 with STxH416 (Orly2)
This commit signifies initial successful testing of U-Boot on a B2000B
board, using STMicroelectronics' STxH416 (Orly2) SoC, including the
ability of the ST40 core being used to boot linux on the ARM core.

To be clear, this U-Boot can boot an ARM linux kernel, although this
U-Boot *only* runs natively on the ST40 (and not the ARM) CPU.

Specifically, the following were tested, and known to work:
	RS-232 Serial Console (ASC0, ASC1, ASC2)
	Ethernet (both MACs) with B2032 (IP1001LF) and B2035 (IP101A)
	Ethernet: 10Mbps (HDX), 100Mbps (FDX), 1000Mbps (FDX).
	EHCI (USB 2.x), on all 4 root ports.
	NAND (Hamming, BCH18, BCH30)
	SPI Serial Flash (FSM, Bit-Banging)
	Boot-from-NAND (Hamming, BCH18, BCH30)
	Boot-from-SPI
	Boot an ARM linux kernel (from ST40)
	I2C

Specifically, the following are untested, or known *not* to work:
	OHCI (USB 1.x)
	The "reset" command
	HoM (Hibernate-on-Memory)
	Boot an ST40 linux kernel (no kernel is currently available)
	B2035 (IP101A) on GMAC #1 (ping works, but nfs fails!)

All the above was tested on a single board: B2000B-00-N033, with a STxH416
(Orly2) cut 1.x in 32-bit mode, with ST40 as the boot-master. Ethernet
was tested using: a B2035A-00-N011 (IP101A), and a B2032B-00-N256
(IP1001LF) board.

Note:	The B2035 (IP101A) worked fine on GMAC #0, but not GMAC #1.
	"ping" worked on both GMACs, but "nfs" only on GMAC #0!

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>