[STM] Completed initial testing of B2000 with STxH415 (Orly1)
This commit signifies initial successful testing of U-Boot on a B2000B
board, using STMicroelectronics' STxH415 (Orly1) SoC, including the
ability of the ST40 core being used to boot linux on the ARM core.

To be clear, this U-Boot can boot an ARM linux kernel, although this
U-Boot *only* runs natively on the ST40 (and not the ARM) CPU.

Specifically, the following were tested, and known to work:
	RS-232 Serial Console (ASC0, ASC1, ASC2)
	Ethernet (both MACs) with B2032 (IP1001LF) and B2035 (IP101A)
	Ethernet: 10Mbps (HDX), 100Mbps (FDX), 1000Mbps (FDX) hubs
	OHCI (USB 1.x) + EHCI (USB 2.x), on 3 root ports
	NAND (Hamming, BCH18, BCH30)
	SPI Serial Flash (FSM, Bit-Banging)
	Boot-from-NAND (Hamming, BCH18, BCH30)
	Boot-from-SPI
	Boot an ARM linux kernel (from ST40)
	I2C

Specifically, the following are untested, or known *not* to work:
	The "reset" command
	HoM (Hibernate-on-Memory)
	Boot an ST40 linux kernel (no defconfig is currently available)

All the above was tested on a single board: B2000B-00-N056, with a STxH415
(Orly1) cut 1.x in 32-bit mode, with ST40 as the boot-master. Ethernet was
tested using: a B2035A-00-N011 (IP101A), and a B2032B-00-N256 (IP1001LF)
board.  Both B2032 and B2035 are reliably working on GMAC #0 and GMAC #1.

Note: NAND was tested using a B2006A, and a pair of B2007A boards -
3 different NAND devices (with page sizes 512, 2KiB, 4KiB).

Note: Although the B2000 has 4 USB ports, the STxH415 (Orly1) only has 3
USB controllers, hence only 3 USB ports can be used. The STxH416 (Orly2)
has 4 USB controllers, thus all 4 USB ports can be used with that SoC.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>