[ST40] Rename the ST40 files into the new arch/... structure.
Following on from the recent re-baselining activity, in particular
the renaming into the new 'arch/...' directory paradigm, this commit
undertakes all the rename work for STMicroelectronics' ST40 tree.

In particular, we rename the following sets of files:

	renamed:    cpu/st40/*         -> arch/st40/cpu/*
	renamed:    include/asm-st40/* -> arch/st40/include/asm/*
	renamed:    lib_st40/config.mk -> arch/st40/config.mk
	renamed:    lib_st40/*         -> arch/st40/lib/*

In addition, the contents of various files were modified to reflect the
new locations of the above files, to keep things consistent.

Testing was limited to building on a handful of ST40 boards.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>