[STM] Complete the merge of 'v2010.09' into STMicroelectronics' tree
The recent commit (28829fd3609ba5b463cbe74119b8977c97801fc9) was a
"snapshot", which did not build cleanly. This commit signifies completion
of the merge process started by that commit. So, we now have a fully
complete merge of the 'v2010.09' release, for both ST40 and SPEAr systems.

This commit effectively, re-bases U-Boot from release 'v2010.06',
to release 'v2010.09', by refactoring (for most of) the myriad of
changes between these two releases, that did not originally result in any
automatic merge conflicts. Basically, this commit re-applies the "essence"
of those changes to the ST40- and SPEAr-specific parts of this tree.

This commit also confirms that all the supported target boards for ST40
are able to be built (i.e. compiles + links).  It should be noted that
this tree compiles, but does *not* link for SPEAr.  The SPEAr linking
issue is due to some arch-specific caching functions (flush_dcache_range,
and invalidate_dcache_range) that were back-ported to be used in EHCI
(USB), but have not yet been back-ported for the ARM.

Note: no testing has been done, modulo compiling+linking as above.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>