[STM] Re-factor "ST40" to "STM" (Part 1)
In preparation for adding the current STMicroelectronics' SoCs containing
ARM cores, we needed to re-factor the code slightly, to tease apart the
pervasive conflation that erroneously, effectively treated "ST40" as an
"alias" for STMicroelectronics. That is, the code often refers to the
(narrow) term "ST40", when (in hindsight) it should have referred to the
(greater) term "STM" (i.e. the greater STMicroelectronics).

This commit changes some of the (mostly conditional) code that, (in
hindsight) is not ST40-specific, but is actually specific to a plurality
of STMicroelectronics SoCs, in general (e.g. ST40, SPEAr and ARM systems).
Hence, the new semantics are:

	CONFIG_STM      = specific to STMicroelectronics (ST40, ARM, SPEAr)
	CONFIG_ARM      = specific to the "ARM" cores
	CONFIG_ST40     = specific to the "ST40" cores
	CONFIG_SPEARxxx = specific to the "SPEAr" cores

In addition, the macro "CONFIG_STM", is now explicitly defined for all
ST40-specific boards (in addition to "CONFIG_ST40" still being defined).

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>