[STM] Re-factor "ST40" to "STM" (Part 2)
In preparation for adding the current STMicroelectronics' SoCs containing
ARM cores, we needed to re-factor the code slightly, to tease apart the
pervasive conflation that erroneously, effectively treated "ST40" as an
"alias" for STMicroelectronics. That is, the code often refers to the
(narrow) term "ST40", when (in hindsight) it should have referred to the
(greater) term "STM" (i.e. the greater STMicroelectronics).

This commit changes some of the (mostly conditional) code that, (in
hindsight) is not ST40-specific, but is actually specific to a plurality
of STMicroelectronics SoCs, in general (e.g. ST40, SPEAr and ARM systems).

Specifically, the following configuration macros were changed:

	CONFIG_ST40_<board>        ->  CONFIG_STM_<board>
	CONFIG_ST40_<soc>          ->  CONFIG_STM_<soc>
	CONFIG_ST40_NAND_*         ->  CONFIG_STM_NAND_*
	CONFIG_ST40_NO_EPLD        ->  CONFIG_STM_NO_EPLD
	CONFIG_ST40_CONTIG_MODE    ->  CONFIG_STM_CONTIG_MODE
	CONFIG_ST40_STM_SATA       ->  CONFIG_STM_SATA
	CONFIG_ST40_STM_USB        ->  CONFIG_STM_USB
	CONFIG_ST40_STX_STX7105_*  ->  CONFIG_STM_STX7105_*
	CONFIG_SYS_ST40_BART_*     ->  CONFIG_SYS_STM_BART_*
	CONFIG_SYS_ST40_NAND_*     ->  CONFIG_SYS_STM_NAND_*

Tested only by compiling all the ST40 boards.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>