[STM] Re-factor "ST40" to "STM" (Part 3)
In preparation for adding the current STMicroelectronics' SoCs containing
ARM cores, we needed to re-factor the code slightly, to tease apart the
pervasive conflation that erroneously, effectively treated "ST40" as an
"alias" for STMicroelectronics. That is, the code often refers to the
(narrow) term "ST40", when (in hindsight) it should have referred to the
(greater) term "STM" (i.e. the greater STMicroelectronics).

This commit removes all references to several macros, all of which are
prefixed with "ST40_". These macros were not specific to the ST40, and
(in hindsight) should have been prefixed with "STM_" instead. However,
whilst reviewing this code, it became apparent that none of these macros
were actually used! Hence, this commit just deletes them all!

	ST40_BIG_ENDIAN
	ST40_LITTLE_ENDIAN
	ST40_CLOCKGEN*
	ST40_DMAC_*
	ST40_EMPI_*
	ST40_ILC_*
	ST40_INTC2_*
	ST40_INTC_*
	ST40_MAILBOX*
	ST40_MPXARB_*
	ST40_NAND_CONTROLLER_VERSION
	ST40_PCI_*
	ST40_SYSCONF_*

This commit should not logically change anything, but it simply eases
maintenance in the future.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>