[STM] Re-factor "ST40" to "STM" (Part 4)
In preparation for adding the current STMicroelectronics' SoCs containing
ARM cores, we needed to re-factor the code slightly, to tease apart the
pervasive conflation that erroneously, effectively treated "ST40" as an
"alias" for STMicroelectronics. That is, the code often refers to the
(narrow) term "ST40", when (in hindsight) it should have referred to the
(greater) term "STM" (i.e. the greater STMicroelectronics).

This commit removes all references to the 2 macros "ST40_LMI_VERSION",
and "ST40_LMIGP_VERSION", which were classed as being ST40-specific, but
were valid for all STMicroelectronics SoCs. However, whilst reviewing this
code, it became apparent that neither of these macros were particularly
useful, as none of the code was logically affected by them!

In addition, all references to the following macros were deleted, as the
macros were only defined in one header file, and never used anywhere else:
	STX5197_SYSSERV_*

This commit should not logically change anything, but it simply eases
maintenance in the future.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>