[STM] Re-factor "ST40" to "STM" (Part 5)
In preparation for adding the current STMicroelectronics' SoCs containing
ARM cores, we needed to re-factor the code slightly, to tease apart the
pervasive conflation that erroneously, effectively treated "ST40" as an
"alias" for STMicroelectronics. That is, the code often refers to the
(narrow) term "ST40", when (in hindsight) it should have referred to the
(greater) term "STM" (i.e. the greater STMicroelectronics).

This commit changes some of the code that, (in hindsight) is
not ST40-specific, but is actually specific to a plurality of
STMicroelectronics SoCs, in general (e.g. ST40, SPEAr and ARM systems).

Specifically, the following macros were changed:

	ST40_ASC*                ->  STM_ASC*
	ST40_EMI_*               ->  STM_EMI_*
	ST40_EMISS_*             ->  STM_EMISS_*
	ST40_LMI*                ->  STM_LMI*
	ST40_PIO*                ->  STM_PIO*
	ST40_SSC*                ->  STM_SSC*

And the following function name was also changed:

	ST40_start_host_control  ->  STM_start_host_control

Tested only by compiling all the ST40 boards.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>