mmc: Define a constant for the maximum block size
The number 512 appears quite a bit in the mmc code. Add a constant for this
so that it can be used here and in other parts of the code (e.g. SPL code
which loads from mmc).

Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Vadim Bendebury <vbendeb@google.com>
(cherry picked from commit 8bfa195e4e7ecbbabd4f070a3640212a5ac4ebde)

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>