[STM] The TMU is specific to the SuperH (and not on ARM)
The "TMU" timer is specific to the SuperH (ST40) cores, and is not
present on ARM systems. As a result, all code that refers to the TMU
should only do so, if we are actually building for a ST40 SoC, and not
for any other STMicroelectronics' SoC.

This commit conditionally compiles the code such that direct references
to the TMU-specific bitfield ("bi_tmu_frq" in "struct bd_info"), and
indirect references through the accessor function "stm_get_tmu_clk_rate()"
will only exist in a valid ST40 context, and not whilst in an ARM context
when compiling for STMicroelectronics SoCs.

Note, for supported SoCs (and boards) that do not have any ARM cores,
no such conditional compilation protection is needed (nor was it added).

Tested only by compiling all the ST40 boards.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>