[STM] Add initial support for ARM on the B2000 board (STxH415/6)
This commit includes initial support for U-Boot (natively) on the ARM
cores in both the STxH415 and STxH416 SoCs from STMicroelectronics. This
support has only been tested on B2000 boards, but it does include
(untested) support for the B2020 for these SoCs as well.

Testing was undertaken only on two B2000 boards, one with a STxH415
(Orly1), and one with a STxH416 (Orly2) fitted, as follows:

	B2000B (B2000B-00-N033) with STxH416 (Orly2) fitted.
	B2000C (B2000C-00-N095) with STxH415 (Orly1) fitted.

Testing comprised: serial console (ASC), Ethernet (B2032/B2035), USB
(EHCI), I2C, and read/write to both NAND and SPI serial flashes.

Testing was run with the instruction caches ON, but with the MMU OFF,
and hence the data caches OFF. All testing was performed exclusively by
downloading U-Boot with GDB (i.e. over JTAG). That is, no testing was
done to boot-from-FLASH, as the "pokeloop" interpreter is still to be
added to support ARM cores for STMicroelectronics.

Still to be done:
	Boot-from-FLASH (NAND and SPI) (i.e. ARM as boot-master)
	Booting ARM linux kernel (no ST40 involved)
	Add support for "Device Trees" (DT).
	And more testing ...

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>