[STM] rename soc_init() to arch_cpu_init() for ST40
The ARM code uses arch_cpu_init(), as the architecture-specific early CPU
initialization function, whereas in the "ST40" tree, we used soc_init().
By renaming the ST40 files to use the ARM convention, we ease the merging
of the two architectures in a unified STMicroelectronics tree.

Tested only by compiling a handful of ST40 boards.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>