[STM] Fix NAND BCH (for ARM) in bch_load_prog_cpu()
In the set of 16 (contiguous) registers, used to describe a single
"Sequence Node", with the BCH controller, (from STMicroelectronics) 2 of
the registers are strictly classed as being "RESERVED".  These reserved
registers are at offsets: +0x21c and +0x228.

Whilst developing code for the ARM core on the STiH416 (Orly2) SoC,
it was noticed that these registers must *not* be accessed. Whereas on
the ST40 core, accessing them appears to be totally benign.

This commit simply ensure that these two registers are not accessed by all
cores (ARM and ST40) on STMicroelectronics SoCs using the BCH controller.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>