[STM] Protect against un-aligned arrays in "auto" variables.
Convert "automatic" (char) arrays (in function scope) to be "static",
to stop them from being copied onto the stack, when the function starts.
This crucially resolves an unaligned access fault, when copying the
initialization string, onto the function's stack, if the string literal
(in '.rodata') is not word (4-byte) aligned.

This problem was initially noticed on the ARMv7, and probably only occurs
if the SCTLR.A bitfield is set to 1, or the MMU is off, and if no valid
"Alignment fault" exception handler is installed. Also, it is only a real
problem if referenced data are stored byte-aligned (and not word-aligned).

Initially, this was only believed to be a problem with initialized
automatic char arrays (i.e. strings on the stack), however, this could
potentially occur with *any* automatic array (of char or short) where
the alignment (of the initializer, nominally in '.rodata') is not a
multiple of 4 bytes.

Making such automatic variables "static", not only avoids a potentially
illegal unaligned access, but it is also more efficient, as there is no
need to copy it onto the stack on every invocation of the function!

So, doing all this is basically: benign, safe, and faster!

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>