[STM] Use recommended STBus bridge values for the GMAC
The recommended STBus bridge values for the STM-GMAC were only correct
for some SoCs, and are wrong for other SoCs. This was because the offset
(was STBUS_BRIDGE_OFFSET), and the value (was STBUS_BRIDGE_MAGIC) used
was the same pair for *all* STMicroelectronics' SoCs.

This commit now uses SoC-specific values for each SoC (where such
values are known).  If the values are not explicitly defined for a
given SoC, then the "poke" to the STBus bridge configuration register
will *not* performed, resulting the "default" configuration (which may
be sub-optimal).

This was noted to be a problem on the STxH416 (Orly2) when developing code
to run on the ARM core, as a poke to the wrong register offset resulted
in a fault, but on the ST40 core the same poke was "benign" (and ignored).

Ideally, the following 2 macros should be defined for each SoC:

	#define STM_GMAC_AHB2STBUS_BASE		<offset>
	#define STM_GMAC_AHB2STBUS_CONFIG	<value>

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>