[STM] Handle unaligned 16-bit fields in the USB stack (for ARM)
The USB stack has some 16-bit fields passed to the USB H/W controller that
are misaligned (i.e. on a 1-byte boundary). These fields are typically
defined in "packed" data-structures, of a specific alignment. This is a
"feature" of the USB specification! The following are such examples:

		wMaxPacketSize
		wHubCharacteristics

Initially, with ARM cores on STMicroelectronics' SoCs, we do not support
any misaligned accesses, so we need to add a few macros to help us, to
force the compiler to convert such potential misaligned accesses into a
series of valid byte accesses, so that we do not generate any code that
could result in a misaligned fault.

These new macros logically perform the following functionality:

	#define UNALIGNED_WRITE16(x,value)	(x) = (value)
	#define UNALIGNED_READ16(x)		(x)
	#define UNALIGNED_OR16(x,value)		(x) |= (value)

On ARM cores, on STMicroelectronics' SoCs, these macros will instead
convert each 16-bit access of these fields, into a corresponding pair
of 8-bit accesses, for the USB stack.

This patch should/could be removed, once STMicroelectronics properly
supports misaligned accesses on ARM cores. It is envisaged that this
would require enabling the MMU as a necessary pre-cursor.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>