[STM] Re-base pokeloop interpreter to version 4.2
This code essentially just updates the STMicroelectronics' "pokeloop"
interpreter to the latest version, which is currently version 4.2.
It also adds support (for the first time) for the pokeloop interpreter
for the ARMv7, in preparation to be used on STMicroelectronics' SoCs.

Essentially, this is just importing the three files: "poketable.h",
"sh4/pokeloop.S" and "armv7/pokeloop.S", from STMicroelectronics'
"poketable.git" repository, which were tagged with "pokeloop_4.2",
into the STMicroelectronics' U-Boot repository.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>