[STM] Boot-from-NAND on B2000 and B2020 boards.
This commit includes initial support for boot-from-NAND flash for
ARM cores, on both the STxH415 (Orly1) and STxH416 (Orly2) SoCs from
STMicroelectronics, on both the B2000 and B2020 boards.

This includes adding support for the logical block zero "magic" signature
pattern that is recognised by STMicroelectronics' BCH NAND controller,
when booting from NAND flash. This allows the system still to boot from
NAND, even when physical block #0 is deemed to be a "bad" block.

In addition, this code will relocate the "poke-table" from FLASH, into
some on-chip SRAM, and it will be this SRAM-based copy of the poke-table,
that will be used by the "pokeloop" interpreter, to perform the low-level
initialization of the clocks, LMI, EMI etc.  This is because whilst the
"pokeloop" interpreter is being executed, both LMI and EMI may temporarily
"disappear" from the CPU, whereas the SRAM should not "disappear"!

It should be noted that only the single designated "boot-master" ARM
core will be executing. Any (and all) slave ARM cores will be placed in
a "holding-pen", and never released, with this code! This clearly needs
to be implemented properly, for SMP usage.

Testing comprised booting U-Boot directly from a NAND flash, using BCH18,
on a single B2000 board with a STxH415 (Orly1) SoC:

	B2000C (B2000C-00-N095) with STxH415 (Orly1) fitted.

Testing was run with the instruction caches ON, but with the MMU OFF,
and hence the data caches OFF.  No testing was performed with a STxH416
(Orly2) SoC, or with a B2020 board, however all these combinations are
expected to work. More testing needs to be undertaken, including BCH30,
and using the Hamming controller.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>