[STM,FIX] Ensure only the correct env_eeprom.c file is used.
Support for U-Boot's environment being stored in EEPROM is normally
implemented in the common file "common/env_eeprom.c". However,
specifically for STMicroelectronics's SoC, we have an additional
STMicroelectronics-specific version of this file, which should totally
usurp the original one: "drivers/stm/env_eeprom.c".  With some recent
"re-naming" activity, of some arch-specific files (specifically commit
6b6060b24ac1593efb87cb863aa54522528f883e), the previously successful means
to include the stm-specific file, over the common one, was inadvertently
broken, and the linker could end up with multiple symbol definitions,
and hence the link could fail.

This commit resolves this problem by ensuring that the make system
will only compile the "common" one, if we are not building for a
STMicroelectronics' SoC. Hence, there should not be any conflicts,
whilst building for STMicroelectronics' SoCs, and the stm-specific file
will always dominate for STMicroelectronics' SoCs.

In addition, the prototype for the function env_init_after_spi_done()
has now been moved into "include/environment.h", and all callers of this
function now #include this header file.

As a corollary, the previous ST40-specific "hack" to pull in the correct
object file, that was in "arch/st40/cpu/start.S" has also been removed.

Tested only by compiling for the B2000 board both with ST40 and ARM.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>