[STM] Boot-from-SPI on B2000 and B2020 boards.
This commit includes initial support for boot-from-SPI serial flash for
ARM cores, on both the STxH415 (Orly1) and STxH416 (Orly2) SoCs from
STMicroelectronics, on both the B2000 and B2020 boards.

It should be noted that only the single designated "boot-master" ARM core
will be executing. Any (and all) slave ARM cores will still be placed
in a "holding-pen", and never released! This clearly still needs to be
implemented properly, for SMP usage.

Testing comprised booting U-Boot directly from a SPI flash (N25Q128)
on a single B2000 board with a STxH415 (Orly1) SoC:

	B2000C (B2000C-00-N095) with STxH415 (Orly1) fitted.

Testing was run with the instruction caches ON, but with the MMU OFF,
and hence the data caches OFF.  No testing was performed with a STxH416
(Orly2) SoC, or with a B2020 board, however all these combinations are
expected to work. More testing needs to be undertaken.

It should be noted that testing was conducted using the SPI S/W
"bit-banging" driver, and not the H/W SPI FSM controller. There is still
an issue to be debugged pertaining to FSM and boot-from-SPI. In the mean
time, please use the S/W "bit-banging" driver as a workaround.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>