[STM,ARM] Relocate the poke interpreter into SRAM to run it
In a recent commit (f760768ba5c169f7d7f96747eafe446418ad2671), support
was added for boot-from-NAND on ARM SoCs from STMicroelectronics. That
commit defined the environment for the poke interpreter (poke_loop()),
such that poke_loop() code was executing out of the I-cache, and only
the poke-table was relocated into SRAM to be used (recall, the D-caches
are still off, as the MMU is off).

However, the offset-data in the "jump-table" in the poke interpreter
was still being fetched from the actual boot-device via the EMI, whilst
the poke interpreter was being executed. This is unsafe, as there should
be zero accesses to the EMI "boot-bank", whilst the poke interpreter is
being executed.

This commit addresses this issue on ARM, by additionally relocating the
poke interpreter, as well as the poke-table into SRAM, and then running
the poke interpreter from SRAM, requiring no accesses to the "boot-bank"
in EMI, whilst the poke interpreter is being executed.

It should be noted that whereas this commit might not be needed for
boot-from-NAND (empirically it worked without it), it is likely that
this commit is mandatory for a reliable boot-from-SPI capability. The
same solution will be used for all ARM-based boot-from-FLASH capabilities.

This code makes some explicit assumptions about the code layout, so it
only needs to copy the minimum data to SRAM - just the poke interpreter,
and the poke-table. These assumptions are:

	1) The poke_loop() (both code+data) is placed immediately before
	   the poke-table (i.e. they abut).

	2) The poke interpreter begins at poke_loop

	3) The poke interpreter is PIC (no fix-ups needed)

	4) The poke-table begins at __memory_setup_table

	5) The poke-table ends at __memory_setup_table_end

	6) Both the poke_loop(), and the poke-table will easily fit in
	   the SRAM, with no issues.

	7) The poke interpreter is entered in Thumb mode

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>