[STM] Initial boot of (native) ARM linux on B2000/B2020
This commit includes initial support for booting an ARM linux kernel
from U-Boot, with U-Boot running on ARM (i.e. a native linux boot).

This includes support for both the STxH415 (Orly1), and the STxH416
(Orly2) SoCs from STMicroelectronics, on both the B2000 and B2020 boards.

It should be noted that this support exclusively supports passing the
boot parameters via the traditional ATAG mechanism. That is, no support
has (yet) been added to support DT (device tress) for linux booting.

There is still no support for SMP linux, as all ARM cores are still held
in the initial "holding-pens". However, a SMP kernel will successfully
boot, but automatically regress to a uni-processor environment.

In addition, the "bdinfo" command has been updated to print two
additional useful values:
	1) The unique Machine ID (i.e. MACH_TYPE_STM_xxxx)
	2) The address used to pass the Boot Parameters (to linux)

Testing comprised booting U-Boot directly from a SPI flash (N25Q128)
on a single B2000 board with a STxH415 (Orly1) SoC:

	B2000C (B2000C-00-N095) with STxH415 (Orly1) fitted.

Testing was run with the instruction caches ON, but with the MMU OFF,
and hence the data caches OFF.  No testing was performed with a STxH416
(Orly2) SoC, or with a B2020 board, however all these combinations are
expected to work. More testing needs to be undertaken.

ADDENDUM: The above testing was conducted by pressing the "reset" button,
and not by power-cycling the board. This is significant, as it later
transpired that booting linux was only possible if GDB has previously
"attached" to the board since the last power-cycle. A future commit will
address this issue.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>