[STM] Support 100% of DRAM on B2000 and B2020 (on ARM)
This commit will use the correct values for both the (physical) address
of the start of the DRAM, and the size of the DRAM, on the B2000 and
B2020 boards. These are as follows:

	Board	Start of RAM	RAM Size
	-----	------------	--------
	B2000	0x60000000	1 GiB
	B2020	0x40000000	2 GiB

The above information was extracted from:
	/opt/STM/STLinux-2.4/devkit/armv7/etc/stlinux_arm_machtypes

On the ST40 cores, the total usable RAM (as seen by U-Boot) is always
capped at a maximum of 512 MiB, due to some architectural decisions made,
pertaining to restrictions using the PMB in U-Boot. Whereas, on ARM,
U-Boot should be able to access 100% of the physical DRAM.

U-Boot itself will reside in the last 1MiB of usable RAM. That is:
	TEXT_BASE = <Start of RAM> + <Usable RAM size> - 1MiB

Testing comprised booting U-Boot directly from a SPI flash (N25Q128)
on a single B2000 board with a STxH415 (Orly1) SoC:

	B2000C (B2000C-00-N095) with STxH415 (Orly1) fitted.

No testing was performed with a STxH416 (Orly2) SoC, or with a B2020
board, however all these combinations are expected to work.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>