[STM] Invalidate the D-caches on ARM, to let linux boot reliably
We need to invalidate the D-caches very early on in the low-level
initialization phases.  Failure to do this, resulted in a U-Boot which
could boot from FLASH successfully, but could not subsequently boot
linux reliably. This commit fixes this, and now allows U-Boot (on ARM)
to boot a native linux reliably from "power-on".

Testing comprised booting U-Boot directly from a SPI flash (N25Q128)
on a single B2000 board with a STxH415 (Orly1) SoC:

	B2000C (B2000C-00-N095) with STxH415 (Orly1) fitted.

Testing was run with the instruction caches ON, but with the MMU OFF,
and hence the data caches OFF.  No testing was performed with a STxH416
(Orly2) SoC, or with a B2020 board, however all these combinations are
expected to work. More testing needs to be undertaken.

There is still no support for SMP linux, as all ARM cores are still held
in the initial "holding-pens". However, a SMP kernel will successfully
boot, but automatically regress to a uni-processor environment.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>