[STM] Correct comments for b2020.h configuration file
This commit is mainly just correcting a few comments in the configuration
file for the B2020 board, from STMicroelectronics.

The original configuration file for the B2020 board was based on the
B2000 one, but a few regions therein were not updated in full. This
commit should address that small over-sight.

Essentially, we just update the comments to reflect:

	* the actual jumper settings (for MODE pins).
	* the component numbers for the (off-board) UARTs
	* the locations for the 4 USB (Type A) connectors.

Also removed references to UART2, as this serial port is not readily
usable on this board.

Testing was limited simply to building for a B2020 board.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>