[STM] Provide better definitions for the SRAM regions
There is now a clear need for two distinct (non-overlapping) "regions"
in SRAM, in U-Boot for ARM SoCs from STMicroelectronics. Hence, the
single macro CONFIG_STM_SRAM_START is not sufficient on its own.

We now add 2 more macros - one for each such region in SRAM:

CONFIG_STM_SRAM_HOLDING_PEN   - Used to store the SRAM "Holding-Pens",
                                and their data.

CONFIG_STM_SRAM_POKE_STORAGE  - Used to store the poke-table, and the
                                poke-interpreter code.

The macro CONFIG_STM_SRAM_START is nominally used to point to the
beginning of the SRAM region, and the other two macros will be defined
(by default) in terms of CONFIG_STM_SRAM_START.

The default values for all 3 are as follows:
    CONFIG_STM_SRAM_START         = 0xc00c0000
    CONFIG_STM_SRAM_HOLDING_PEN   = (CONFIG_STM_SRAM_START + 0x0000)
    CONFIG_STM_SRAM_POKE_STORAGE  = (CONFIG_STM_SRAM_START + 0x0100)

Any and all of these 3 macros may be defined in the board-specific
configuration file, and those values therein will dominate.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>