[STM] Add initial support for SRAM "Holding-Pens" for ARM
This commit adds initial working support for "Holding-Pens", resident
in SRAM, for ARM slave cores, in SoCs for STMicroelectronics.

The idea is that the boot-master core (running U-Boot) will copy (from
FLASH) to SRAM some Holding-Pen (code+data), and then "kick" (with "SEV")
each slave to transition from the initial FLASH Holding-Pen into the
newly created SRAM Holding-Pen. Clearly, this assumes we can reserve
some memory in SRAM for this exclusive purpose.

We introduce the concept of a "jump pointer", which is a (4-byte) pointer,
also stored in SRAM, which yields the address the slave is to "jump" to,
when the slave is "kicked" with a "SEV" instruction from the boot-master
core. We initially store the special sentinel HPEN_KEEP_LOOPING_VALUE in
the "jump pointer", which has the semantics to "keep looping", and not to
"jump" (yet). For (power) efficiency reasons, the slave cores will wait
using a "WFE" instruction, and only re-read & re-test the contents of the
"jump pointer" on receipt of an event.

The boot-master core will try and "herd" (or marshal) all the slave
cores into the SRAM Holding-Pen very early on - specifically before the
"poke-loop" interpreter (i.e. init_ram()) is invoked.  All the slave
cores should then remain in their SRAM Holding-Pens, until such time
as U-Boot is ready to pass control to the next "payload" application
(typically linux).

This commit is limited only to the first half of this process - that is,
just the transition into the SRAM Holding-Pens, and not the transition to
the next "payload" application. Hence, once the slaves are in the SRAM
Holding-Pen, then they will be suck there forever with this commit! A
subsequent commit will address this blatant deficiency.

This was tested on a single B2020 board, booted directly from SPI. Once
booted, GDB was connected to the slave core, and verified that the core
was in a tight-loop, running in SRAM, spinning on "WFE".

	B2020 (B2020C-02-R247), with STxH415 (Orly1, cut 2.x) fitted.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>