[ARM] Invalidate the Branch Predictor, before setting SCTLR.Z
Whilst preparing to add code to handle the "Holding Pens" for the ARM SoCs
specifically from STMicroelectronics, it was noted that random "hangs"
were occasionally observed. Further investigation indicated that if the
Branch Predictor was never enabled (i.e. SCTLR.Z=0), then the observed
problem was not present. This commit is an attempt to solve this issue
by explicitly invaliding the Branch Predictor array, prior to enabling
the Branch Predictor (i.e. SCTLR.Z=1).

This commit just adds code to invalidate the Branch Predictor array,
which we should do before enabling the Branch Predictor.  In addition we
perform a pair of synchronization barriers, to complete the invalidate
sequence safely.

It should be noted that the 3 lines added were present in, and hence
copied from commit c2dd0d45540397704de9b13287417d21049d34c6. Thanks to
Aneesh V <aneesh@ti.com> for the original code.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>