[STM] Disable the Branch Predictor on STMicroelectronics' SoCs
Whilst preparing to add code to handle the "Holding Pens" for the ARM
SoCs, specifically from STMicroelectronics, it was noted that random
"hangs" were occasionally observed. Following on from the previous
investigation (see commit c7d1b067d989f5a6fae7d9b5f63eb64c36a6932b),
despite that apparent fix, the Branch Predictor (BP) was still
demonstrating random "hangs" when enabled, but no such issues when it
was disabled.

If the Branch Predictor was never enabled (i.e. SCTLR.Z=0), then the
observed problem was never observed. Thus, although the aforementioned
commit *should* have fixed this issue, it looks like there is some other
unknown factor involved! It is suspected that perhaps the BP assumes
the MMU is always enabled, if the BP is enabled - this is not true in
the current version of U-Boot for the STMicroelectronics' SoCs.

As a short-term "fix", this commit will simply disable the BP capability
for STMicroelectronics' SoCs, until it can be investigated fully later.
Hopefully, this patch can be fully reversed, once the MMU is has been
enabled for these SoCs.

With this fix, 100% reliable - without it, hangs occur ~ 50% of the time.
Note: first ran "make romgen" to generate the poke-table used.

Tested on one board (with boot-from-SPI):
	B2020 (B2020C-02-R247), with STxH415 (Orly1, cut 2.x) fitted.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>