[STM] Move ATAGS not to conflict with linux's secondary entry-point mechanism.
Prior to this commit, the address of the ATAGS (bd->bi_boot_params),
used by U-Boot, was hard-wired to be CONFIG_SYS_SDRAM_BASE + 0x100,
that is 256 Bytes above the physical start of SD-RAM.

However, the linux kernel needs to convey to its boot-loader, the
physical address of the secondary entry point for the slave cores. This
information is (currently) hard-wired in linux to be passed via the
address CONFIG_PAGE_OFFSET + 0x200.

As the ATAGS can easily be (and usually are) more than 256 bytes, and
as the physical address of CONFIG_PAGE_OFFSET can be the same physical
address as CONFIG_SYS_SDRAM_BASE, then there is a huge likelihood of a
horrible conflict!

To resolve this, we increase the offset of the start of the ATAGS
slightly (from +0x100 to +0x220). In addition, we will change the base
address from the (static) start of SD-RAM, to the (dynamic) entry point
provided in the ".ub" image header. That is, the actual address of the
ATAGS will now depend on how the linux kernel was linked, and how the
".ub" file was built.

The key point, is that there should no longer be a conflict of addresses.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>