[STM] Add support for SMP on linux on SoCs from STMicroelectronics
Following on from the penultimate commit, this commit completes the
multi-stage "Holding-Pen" mechanism, such that we can now "herd" (or
marshal) all the slave cores into linux, and hence boot linux *with* SMP!

Essentially, this commit just exports (and calls) a new (assembler)
function "prepare_hpen_for_linux()", which will prepare all the slave
cores to exit from U-Boot's SRAM Holding-Pen, and subsequently enter
into the linux kernel's secondary entry point "stm_secondary_startup()".
We achieve this by adding a second "jump pointer" in SRAM, dedicated
to dealing with passing control to the next payload application -
specifically the linux kernel!

It should be noted that (currently) the linux kernel does *not* use the
"SEV" instruction to generate an event to "kick" the slaves, instead it
sends a S/W interrupt to each slave core in turn. This means that our SRAM
Holding-Pen code should ideally use a "WFI", and not a "WFE". However,
I was able to get this to work with a "WFE", but *not* with a "WFI"!
I suspect I am just missing a few "pokes" to route the S/W interrupt
correctly on the slave cores.  However, using "WFE" does empirically work,
and probably should be safe.  Should it be required to get "WFI" working
in the future, I have left some vestigial development code for ICCICR,
which is probably necessary, but clearly not sufficient to use "WFI".

This was tested on a single B2020 board, booted directly from SPI.

	B2020 (B2020C-02-R247), with STxH415 (Orly1, cut 2.x) fitted.

Once booted, an appropriate SMP linux kernel was copied (from NAND),
and booted. The console log messages confirmed SMP was activated:

[    0.180000] Brought up 2 CPUs
[    0.180000] SMP: Total of 2 processors activated (3981.31 BogoMIPS).

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>